requirements.txt
Flask==3.0.0
numpy==1.26.2
Pillow==10.1.0
python-dotenv>=1.0.1
scikit-image==0.24.0
scikit-learn==1.5.0
fastapi==0.114.0
pandas==2.2.2
tensorflow==2.15.0
keras==2.15.0
gunicorn
